Phase 2 – ML Prediction Pipeline/requirements_serverless.txt
pandas
pyyaml
nba_api
